requirements.txt
fastapi
uvicorn
pandas
matplotlib
numpy
python-dotenv
aiofiles
langchain
openai
chromadb
langchain-community
langchain-openai
